todoをアカウントで取得するサービスの作成

+++ todo-domain/src/main/java/com/example/todo/domain/model/FindByAccountIdForTodo.java
@@ -20,6 +20,6 @@ public interface FindByAccountIdForTodo {
 	/**
 	 * @see Pageable#getOffset()
 	 */
-	long getOffset();
+	int getOffset();
 	
 }
