+++ todo-domain/src/test/java/com/example/todo/domain/repository/todo/TodoRepositoryTest.java
@@ -153,7 +153,7 @@ public class TodoRepositoryTest {
 		}
 
 		@Override
-		public long getOffset() {
+		public int getOffset() {
 			return offset;
 		}
